+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 2014.04.23
+* Converted Releaf::Node to Releaf::ContnetNode module.
+
+  Instread of inheriting from Releaf::Node, inherit from ActiveRecord::Base and
+  include Releaf::ContentNode
+
 ### 2014.04.22
 * Releaf::Node is in refactoring process. The goal is to make it an abstract
   model (some day)